requirements.txt
pytest
allure-pytest
faker
pytest-xdist
pytest-playwright
playwright
PyAutoGUI
allure-python-commons
requests